+++ iptocc/__init__.py
@@ -3,10 +3,8 @@ import ipaddress
 import os
 from functools import lru_cache
 
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
+import unqlite
 
-from iptocc.models import Record, Base
 
 __author__ = "Ronie Martinez"
 __copyright__ = "Copyright 2017, Ronie Martinez"
@@ -18,26 +16,26 @@ __email__ = "ronmarti18@gmail.com"
 __status__ = "Production"
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
-engine = create_engine("sqlite:///{}".format(os.path.join(dir_path, 'rir_statistics_exchange.db')))
-Session = sessionmaker(bind=engine)
-session = Session()
+database = unqlite.UnQLite(os.path.join(dir_path, 'rir_statistics_exchange.db'))
+ipv4_all = database.collection('ipv4').all()
+ipv6_all = database.collection('ipv6').all()
 
 
 @lru_cache(maxsize=100000)
 def ipv4_get_country_code(ip_address):
-    for record in session.query(Record).filter_by(type='ipv4'):
-        start_address = ipaddress.IPv4Address(record.start)
-        if start_address <= ip_address < start_address + record.value:
-            return record.country_code
+    for record in ipv4_all:
+        start_address = ipaddress.IPv4Address(record.get('start'))
+        if start_address <= ip_address < start_address + record.get('value'):
+            return record.get('country_code')
     return None
 
 
 @lru_cache(maxsize=100000)
 def ipv6_get_country_code(ip_address):
-    for record in session.query(Record).filter_by(type='ipv6'):
-        network = ipaddress.IPv6Network('{}/{}'.format(record.start, record.value))
+    for record in ipv6_all:
+        network = ipaddress.IPv6Network('{}/{}'.format(record.get('start'), record.get('value')))
         if ip_address in network:
-            return record.country_code
+            return record.get('country_code')
     return None