+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Changes to IPToCC.
 
+## [Unreleased]
+
+## [v1.0.3]
+### Changed
+- Used unqlite-python instead of SQLAlchemy(sqlite) for thread-safety.
+- Used ftp://ftp.afrinic.net/stats/afrinic/delegated-afrinic-extended-latest.
+
 ## [v1.0.2]
 ### Fixed
 - Cannot add subdirs in MANIFEST.
@@ -11,6 +18,7 @@ Changes to IPToCC.
 - Support for IPv4 and IPv6 addresses.
 - PyPi setup.
 
-[Unreleased]: https://github.com/Code-ReaQtor/IPToCC/compare/v1.0.2...master
+[Unreleased]: https://github.com/Code-ReaQtor/IPToCC/compare/1.0.3...master
+[v1.0.3]: https://github.com/Code-ReaQtor/IPToCC/releases/tag/1.0.3
 [v1.0.2]: https://github.com/Code-ReaQtor/IPToCC/releases/tag/1.0.2
 [v1.0.1]: https://github.com/Code-ReaQtor/IPToCC/releases/tag/1.0.1